Fix resetting Crazy Tracy message state when leaving and base price on actual kills.
While it was never a huge deal, if you talk to tracy, see her initial message, then leave, next time you enter the building, she will pick up where she left off. So if you have a potion, talk to her, leave, get hit and lose your potion, then go back and talk to her, she will say "But I'm not selling it to you" even when you need a potion. This fixes that, and bases the price on real kills instead of piece of power counter, which is a slight accuracy increase. Like hitting piece of power at 44 kills resets to 0, so there is two continuous kill states where the price is lower and constantly fluctuates.