Allow borrowing the rooster if "Nothing is Missable" is enabled and level 8 is completed.
This is the first step in "borrowing" the followers at the end of the game, giving one more chance to make use of them before finishing the game. The rooster allows accessing a seashell that could be missed near Kanalet Castle. With the redux option enabled, he can be borrowed from the hen house and returned when finished.