Marin will take a walk if "Nothing is Missable" is enabled and level 8 is completed.
This was nothing short of a nightmare to get working correctly. The rooster was easy because it is inside a house, and a map transition takes place shortly after. But Marin is out in the open, and this "engine" is... I can't even put words to it. Nothing works like you think it does, or just doesn't work at all. I could go into so many fine details, but this isn't the place for that.