Fix the exit direction of several caves and passageways to not feel awkward.
Several of the cave and stairway exits around Koholint Island felt weird in that, it would push you in the opposite direction that you would want to go. For example, the passageway entering Tal Tal Mountain Range. You enter the cave right of the egg, push a few rocks, take the stairs, and when exiting those stairs, it pushes you left a bit when you need to go right. A few other caves where this was fixed is the passageway to animal village, the passageway to the upgrade bat in Martha's Bay, and a few caves on the mountain.