Play a jingle and stop the music when meeting Marin at the beach.
When approaching the "screen" where Marin is standing along the shore, in the original game a happy little jingle plays then the music stops and all you can hear is the waves. When leaving the screen and not talking to Marin, the music resumes. This is now exactly what happens in this port.