Another "hookshot is stuck and wont' fire" fix, except this time it's fool proof.
A little while ago I found the hookshot could get stuck if interrupting it with another weapon. This was really easy to do, so I implemented a quick fix for it: https://github.com/BigheadSMZ/Zelda-LA-DX-HD-Updated/commit/4d2fb368c3a98a7b0e2e9141857e48b5d902ac47

Looks like there are other ways to get stuck, like pulling in a piece of power or acorn at close range. The hookshot is instantly despawned and is never "returned". So to fix this once and for all, a countdown is used to check if the hookshot has returned yet. If it hasn't, reset the firing variable. Easy.