In level 5 dungeon, playing Ballad of the Windfish forces Gohma's eyes open.
For all the years I've been playing this game, I didn't know this secret existed until I seen someone mention on itch.io that this didn't work in this port. I could not find any videos on it, and I'm too lazy to play up to that point on GBC so this implementation may not be 100% correct. But for now, it's better than no implementation at all.