No more picking up invisible roosters in level 7.
How funny is it that I went to pick up the ball and all of a sudden I'm flying around with the rooster in level 7? Not very funny at all if you're me. Turns out the rooster is always with you until you finish level 7, just invisible and can't be interacted with. Well at least that's how it's supposed to be. The fix in 1.8 to not let picking up the dead bones allowed picking him up in places he shouldn't exist. Hopefully this is fixed now...