Fix multiple issues with the final boss. Giant Zol form, bombs vs. DethI, and more.
For some reason I have a blind spot when it comes to the last boss, maybe because it's a pain to test and battle through all the various states. But I'm at a point where it needs to be done. This really should have been in v1.2.2 so I'm going to release an update to v1.2.3 much earlier than I usually do.

This fixes Giant Zol form from getting stuck in the "elongated" state. It also changes the odds the Agahnim state will fire the exploding shot from 25% to 33%. The music will now properly stop after defeating the Lanmola state. DethI can now be damaged by bombs. And finally, disable the low heart alarm once the final form is extinguished so it's not beeping all through the ending.