Level 1 dungeon: Don't let Keese outside the room required to defeat them to continue.
So this was funny. Was messing around and a Keese flew outside the wall and the doors close behind you when entering this room so there was no way to kill it. So, I made Keese unable to go outside of "enemy walls" which are walls that enemies can not cross if they have the proper flag set. Then I surrounded the room with these walls. Problem solved forever.