Force Link back into an "idle" state when encountering the owl.
Because of the code I implemented to freeze everything around the map when the owl shows up, it had the nasty side effect of freezing Link in whatever state he was in as the Owl approaches. So instead of that, return him back to his "idle" state so the owl isn't talking to a frozen jumping link with a sword pointing in some direction. This is not a "clean" transition as it's instant, but it beats the former behavior.