Implement the secret ending. It was already kind of in the game, but was very unfinished.
This implements the full "secret ending" that appears after the credits when completing the game with 0 deaths. It can also be seen in this port if you enable "Nothing is MIssable" in the Redux options. This ending was present, but it was very broken and very incomplete. Basically all it did was show a portrait and send you back to the intro. No music, no animations, no fade effects, nothing. So I implemented everything that is supposed to happen during this ending.