Gohma: Play Ballad of Windfish - Eye opens 1.5 seconds - No fireball shot.
I still don't know if this is accurate behavior, would love to see a video of this or something. Maybe one of these days I'll make it up to this point in the original game again just to see what is correct here. How it was previously was kind of annoying. The eye opened but a fireball was always shot which isn't what I remember. But maybe I'm wrong.