Add enemy wall between north/south face shrines and make bomber react to it.
This helps with the mushroom bomber messing with you when making your way to the southern face shrine. It was not there in the original game because it couldn't cross over the screen. The added invisible "enemy wall" helps with this, but it's not a perfect solution as it can still slip the occasional bomb through. But with this, it's much improved from before.