Fix the "Sound Effects" volume slider which had no effect.
Sadly, I broke this a looooong time ago when fixing the "Mute Inactive Window" option. Unfortunately, it broke the volume effects slider in the process and it took me a long time to notice. Then it took me even longer to fix. Sadly, this is the first time both "Mute Inactive Window" and "Sound Effects Volume" both work together.