Finally fix Marin sometimes not spawning as a follower after the beach scene.
This was a nightmare manifested into reality. Troubleshooting this involved many long hours of writing diagnostic code, testing, followed by a whole bunch of disappointment. The actual "bug" was never discovered. The "Marin spawner" for some reason doesn't work when playing from the start of a new file up to the point of the beach scene, and without saving/quitting/continuing and/or getting a game over which reloads the save. The workaround is to just spawn her directly and don't rely on the Marin spawner. We may never know the real cause of why this doesn't work like it should. But as long as it works with this method, who cares?