Add missing sound effects to both Vire's fireball attack and fireball death.
Vire is the bat enemy in Level 8 that flies around the room spitting fireballs at you and splits up into two bats when hit. This is an enemy that's been around since the very first "The Legend of Zelda" game on the NES. I noticed a "ToDo" comment and then checked it out and yeah, it had no sound effects.