Sword no longer blocks projectiles but add modifier to do this.
Sword never should have been able to block projectiles despite it being somewhat fun to do so. So I've added a modifier to do it. But the modifier only works when swinging the sword. Holding it out in front of you no longer blocks projectiles as it kind of invalidates the purpose of the shield.