Fix Boomerang to deal damage regardless of current Z position.
The Boomerang was designed in a strange way compared to other projectiles. Unlike most, it ignores Z collision for travel but is not ignored for hit detection. As it travels, it gains +4 to the Z position over time starting from 0, no matter the starting height of the player. So when standing on blocks that have a depth of 8, it technically travels "under" enemies. To make sure it always hits everything on any layer, I gave it a steep depth of 20.