Add "Save & Continue" option & rename "Quit to Menu" to "Save & Quit to Menu".
This gives the option to save the game on the spot and keep on playing. If "auto-save" was not enabled, I don't think there was another way to save other than to just quit the current gameplay session. This also makes it more clear that quitting to the menu will also save the game.