Add option to lock the camera so the right joystick has no effect.
This "disables" the free camera, default enabled since I always make options to go beyond what is "normal" when compared to the original game. In other words, the original game did not have a free camera option, so the default state should be disabled. But like all these options, set it once and it's on until you turn it off.