Completely disable interacting with Bosses upon their defeat.
This disables the "PushableComponent" and the "HittableComponent" from bosses when they are defeated, making it impossible to interact with them as they enter the "dying" phase. Some bosses could still be smacked around with the sword or bumped with the shield. You had your fun, but it's time to pack it up and go home.