Make charged sword attack spin clockwise when facing right.
This is one I've had on my radar for awhile, but I knew it would be some work since it would mean updating both the sword and Link animation files. But, it wasn't too bad, just took some time to get perfect. Anyway, in the original games, Link's sword spin is counter clockwise when facing up, left, and down. But, when facing right, the sword spin is clockwise. For the sake of accuracy, this has now been fixed.