Spawn a "sprite" shadow under Link when generated shadows are disabled.
This adds a shadow under Link using a sprite instead of a dynamically generated shadow when shadows are disabled. I tried to make it versatile enough that it can also be added to other objects. Just create a new shadow, feed the host object to the parameters along with an offset, and the rest should take care of itself. While this is simple enough, it will still take quite a bit of work figuring out what all objects will need a shadow, and actually adding it to each of those objects.

This should also benefit systems that have trouble rendering dynamic shadows. I guess Winlator, a Windows emulator for Android has trouble generating dynamic shadows for whatever reason, but since it can render sprites it should be able to render a sprite shadow. The shadow sprite was added to "objects.png" so a patch can be applied without needing additional files.