Modify Hardhat Beetle movements to more closely reflect the original games.
The way it was working, was that when the beetle was initialized, it had a random movement speed set which never changed until death and a new one spawned. This was not correct behavior, and was most likely done to prevent "sprite stacking". The way it works now, is every 3/4 of a second they may gain a slight "speed boost" which is how it seems to work in the original games. It's probably not perfect, but it's much closer than the previous way it was done.