Fix Hookshot crash when grabbing fairy and fix Boomerang to carry fairy instead of instant collect.
Hookshot crashed when trying to grab a fairy since it had no idea what the object was, so add the same fix as with the Boomerang. And since I'm here, the Boomerang collecting a fairy was something I did early on when I was still a total noob and had no idea how things worked. Now I do understand everything much better, so let's fix it properly now. Boomerang used to "instantly" collect the fairy on contact, now it along with the Hookshot, will properly carry the fairy back to Link.