Animal Damage: Added Redux option - Enabled: no hits - Disabled: burning allowed.
So I've had requests for both of the extremes over time: an option to not allow damaging animals at all, and also restoring being able to set them on fire with powder/magic rod like in the original games. My goal is always to get as close to the original games as possible (within reason),  so the ability to make fried chicken has been restored. I've also added a "Redux" option to completely disable animal damage altogether, which can be useful if you don't want to accidentally attack dogs and get hit. However you want to play, the world is your oyster.. whatever that means.