Explicitly disable followers on certain maps + fix Marin dungeon sequence.
This change does a few things. First, it enables/disables followers based on specific maps rather than trying to rely on some archaic code somewhere hidden to do it. Second, it fixes the Marin "dungeon sequence" when leaving a dungeon that has been broken since the change that fixes her beach spawn. Third, it adds this sequence to many more dungeons. The full list includes: levels 1, 2, 3, 4, 7, and 8. Level 5 is in the water so nah, and Level 6 entrance isn't spawned in right away. For those dungeons, followers just won't appear.