Completely rework shield hitbox/mechanics and remove enemy components on death.
This ended up being a rather enormous change, spread across multiple stages all relating to how the shield works mechanically.
- Rework shield collision boxes. Modify based on facing direction.
- Calculate projectile direction based on angle rather than vectors.
- Rework hit detection of enemy body box colliding with player body box.
- Always block if projectile is in shield box even if it's also in body hitbox.
- Assign enemy components to fields so they can be deactivated on death.

Overall, this fixes some fundamental flaws with how the shield worked. If dashing at an enemy with sword + shield out, the player could sometimes get hit because the projectile would "skip" the shield hitbox and make it to the body hitbox. Now it always blocks if the projectile touched the shield hitbox. The shield would also bounce off the dying enemy after it was killed if moving fast enough, which had no "quick fix" that didn't have drawbacks. The best approach, and the most efficient, was to deactivate the components of the dying enemies.