Implement method to load alternative language lines.
This was done at the request of MasterPhW who is working on the upcoming German translation. This adds an option to load alternative lines from "dialog_{language}.lng", allowing for multiple translations in a single language file. Currently supports: Original Text, Modern Remake, and Alternative. It will be up to the community to add additional translations to the existing language files.

To add additional lines, we use suffixes "_mod" and "_alt". Assume a line that starts with the key "text_line" and has the message "I'm a line of text". To add an alternative to this, you would use "text_line_alt" and then a new translation like "I'm a line made up of text". Then if the Switch remake had something different, you do "text_line_mod" and "I am a line made up of text". Each line can have up to three different translations, that shows up depending on the option state.