Because nothing is ever simple, fix issues with loading language textures.
Trying to get around languages being tied to an index has not been an easy thing to defeat. The problem with an index based approach is that the index can change depending on the number of languages. Lets say we have english, spanish (esp), and french (fre). English is always index 0, but the next one is spanish (esp) so its index is 1, last is french (fre) so index 2. Now later, we add german (deu), this gets index 1, spanish index 2, and french index 3. You can see how this can be a problem when referencing them by index. The number of languages must always be static. Hopefully with this change, we index based language references are defeated forever.