Fix screen going black if hitbox + damagebox were exact same position.
This was a weird one. The direction of incoming damage could cause render targets to fail if both damage box and player hit box were perfectly aligned and the vector ended up at 0,0. Calling "normalize" on the vector resulted in a division by zero, which made "direction" not a number "NaN" causing the entire screen to go black. Fun stuff.