Reload photo book sprites when refreshing dynamic resources.
The actual photographs were refreshed on language change and when changing the "Colored Photos" settings, but not the supporting sprites. This used to work correctly when it was just the Spanish textures, but due to some much larger, more complicating changes to support all language textures, the game now defaults to English on boot and refreshes dynamic sprites later on.