Fix overworld teleporters indexes + increment save file version.
If all teleporters are discovered, the teleport order should be:
1.) Teleporter outside Mabe Village.
2.) Teleporter near level 8 back entrances.
3.) Teleporter between level 4 and raft shop.
4.) Teleporter near animal village.

Teleporters 2 and 3 were reversed and this commit fixes the teleport order. Because the indexes have been swapped, it also swaps which ones are currently discovered. Version 1 save files are modified to fix the discovered indexes on the fly, and increments the save file version to version 2. This commit also moves setting the current save file version  on new game creation to "GameManager" instead of setting it in "script.zScript".