Don't let player forcefully pass the egg boundary with followers.
If running towards the egg with boots with a follower, it was possible to break past the boundary put in place to prevent followers from entering the egg. The issue was a cooldown was imposed on the boundary, so reduce it lower than the actual "automatic control" time. Link also had a sword animation glitch, so stop running when an auto command is issues.