Only check for followers when going to egg after level 8 has been completed.
We don't need to check for followers before that since they can't get into the egg until it's open. Under normal circumstances, the player will not have a follower by the time the egg can be opened. And it's not until after level 8 that "Nothing is Missable" allows getting followers back.

This also adds a dialog for the rooster when trying to enter (just a squawk) and fixes any weird animations when a turnaround is forced.