Add Redux option to enable skipping dialog boxes with Start button.
This is a much requested feature finally implemented. Consider it experimental at the moment as I don't know if it will cause any issues. From my testing it seems fine. The thing I worry about the most is dialogs that have a "Yes/No" option. Which answer it defaults to is uncertain to me.