Fix room in Dungeon 8 that is supposed to be dark. Also make dark rooms darker.
There is a room with two Ropes (the snake), a peahat (flying flower), two torches, and some blocks that was fully lit that should have been a dark room. This fixes it so the room is dark when entering and isn't lit unless the torches are lit. This commit also makes dark rooms darker so they appear closer to the darkness of the original games.