Support custom values for "ObjDungeonBlacker" via mod support.
Despite the name, this object controls lighting for nearly every map in the game aside from "overworld.map". It allows controlling four lighting values: Red, Green, Blue, Alpha; where alpha is how "bright" the light is. This is global across the entire map, and does not affect "dark rooms" found in dungeons which is a separate object.