Play Marin "idle" animation when facing forward.
All this time Marin should have had an animation when facing forward. The "stand" animation is used when she is idle, but the "idle" animation is the one that is animated. We can cycle between these animations based on the direction she is facing, so when facing forward, play "idle". When she is facing any other direction, play "stand".