Make sword hit objects with Z axis from -8 to +8 relative to position.
This makes it so the sword is able to hit both "above" and "below" wherever Link is currently positioned relative to the Z plane by 8 unit. For example, if he's at default 0, he could hit objects that are at -8 up to +8. If he is standing on those barriers that raise which puts him at 4, he could hit -4 up to +12. This should fix any "height" issues where the sword doesn't connect that did not exist in the original games.