Allow collecting Guardian Acorn & Piece of Power when jumping/slashing.
Also slightly increases the height of the hitboxes of the rupee and acorn so that their hitbox matches their actual sprite size. All other items will either match their actual size, or be reduced to at least 12 pixels in height. This change lets the sword collect acorns and triforces in situations where an enemy is killed, it flies beyond the boundaries of the room up onto the wall, and the only way to reach it is jumping and sword slashing. This worked in the original games so it should now work here too.