Fix playing the landing sound effect when jumping on cracked floors.
This one took me down a rabbit hole, but at least I managed to get out of Wonderland by the end of it all. It seems I don't understand how "GetComponentList" actually works, or it works in a way I don't understand. Replacing it with "GetObjectsOfType" which also takes a rectangle ended up being a better choice, but still only half the battle. The second part of this is that cracked floors have a "ObjHole" associated with each of them, just inactive until the floor is broken. So the hole is there, you just can't see it, which is why the sound effect was not playing. Well, half of the reason why it wasn't playing.