Unleash the flower power. Fix the two instances of swamp flowers to be animated.
There are two patches of yellow flowers in the swamp. These flowers had the frames available, an entry in the sprite atlas, but were never actually added as a game object and never added to the map. The flowers found around the map have two versions: a tile, and a game object. The tile merely serves as a placeholder to know where to place the game object, as only the game object is animated. So to wrap up this overly long description for a simple change, the game object version of the swamp flowers was created and placed over the tile version. This change also fixes the animation speed of all flowers to be closer to the original game.