Support additional movement options via "ObjLink.lahdmod".
This makes it possible to further modify the way Link moves in-game. Currently there is the "Modern Controls" which is default, and the "Classic Controls" which makes movement identical to the original games. This mod file allows further configuration where the analog stick can have "modern" controls while the DPad has "classic" controls. It's also possible to disable the "moonwalk" which was present in the original games control scheme, which was possible to trigger by pressing a diagonal input in the opposite direction Link is facing.