Marin Follower: Play proper sound when jumping off the side of a cliff.
There is a different sound for when a regular jump takes place vs. a cliff side jump. This port played the regular jump sound for both for Marin. This seemed a bit weird especially following Link whose jump does play the correct sound, so fix it so both Link and Marin's cliff jump sound effects are the same.