Add "lahdmod" file to configure enemy lives individually.
The in-game option "Extra Enemy HP" is a pretty generic option in that it increases the amount of HP for all enemies by the set amount. This mod file allows configuring each monster individually. The values here serves as the "base" value, so when using the Extra Enemy HP setting, the final value is the HP set in this file + the value added by Extra Enemy HP.