Make holes even more accurate to original games.
While the last hole change was close, I still wasn't entirely happy with the result. So I turned to AI to try to help me figure out a weird quirk that the original games have. And that is the bottom half of the hole pulls "later" than the top half of the hole. This could be due to a plethora of differences, bodybox size, actual hole boxsize, the amount of "pull", who knows. So to get it 100% accurate is most likely a pipedream. But, we can sure get it close.

This change makes when holes start pulling on both the top half and the bottom half accurate to sub-pixel precision. You can stand just as close to the hole as you can in the original games. And as soon as you cross that threshold, the hole will swallow you up. There is probably still (some) differences, but I can not reasonably get it any closer than this.