Move NPC collision type enum to bottom of the list.
Once again I'm left baffled and wondering what exactly is going on here... Several enums use a name that points to an integer that is usually used to represent stuff like memory regions. I figured it's just whatever, that as long as each number is unique it doesn't matter. Boy was I wrong, as squeezing in "NPC" between two values broke the camera mouse sequence. Giving Link the NPC collision caused him to infinitely fall down holes. How? NO CLUE as ALWAYS HAHAHA. There is nowhere in code that specific bitmasks are checked for, nor the index of the enum, so why it works this way is a mystery to me.