Update objects to use "ModFile.Parse" method + add support for additional "lighting" files.
This updates objects to use the new "ModFile" class to use it's "Parse" method to reduce code bloat. It also adds support for a few more objects that have light sources built into them. The point is to be able to create a "modpack" that restores "classic lighting", or rather, no lighting at all like this port has. This batch of lahdmod files should be enough to get the job done.