Fix Level 7 barrier (again) this time when pushing block to the right of it.
This is the 2nd attempt at fixing this since my last one wasn't good enough. Instead of just removing the first barrier found when pushing the block, now the direction is checked and the position of the barrier is verified. The block must be pushed up, and the position must be exactly the first barrier in that row.