Create tracking variable to reference when a game has started.
There isn't a simple way to know if a file selection has been made and the game is past the title screen. This creates a variable that does just that. When a save file is loaded or a new game is started, set "InProgress" to true. When the player quits the game and goes back to title screen, "InProgress" is set to false. So far this isn't really used anywhere, but there were times when I thought it could be useful so it's available now at least.