Classic Camera: Implement global field on ObjLink and prevent other items from escaping it.
This commit puts the field on ObjLink so it only ever needs to be calculated once every frame which all objects can reference rather than needing to calculate it all over the codebase. The boomerang and hookshot have been updated to reference this field, and the magic rod, sword shot, and arrow are now contained to the current field and collide with the screen edges.