Classic Camera: Ensure camera remains pixel accurate through screenshake & transitions.
This was one of the toughest problems to solve, and there may still be situations that throw the camera off. Basically, when in "classic camera" mode, the camera should "snap" to the same field rectangles as the original game. A pixel in this port is not the same as the original game, one "pixel" could be 10x10 pixels when scaled by 10x for example. Things like screen shaking or map transitions that force the camera to recenter could throw off the "center point" of the camera relative to the field by a few subpixels (or rather, real pixels). This is due to everything being tracked with floats, where we may get decimal values like 3.9999~ instead of 4 for example. This commit tries to ensure we are always working with whole numbers by rounding the results. Even if somehow the camera gets offset in some other way, a screen shake should snap it back to the proper field coordinates.