Fix the inventory + map overlays to correctly scale with UI scale.
Unfortunately, I wasn't able to give up until this issue was solved despite the nightmare involved in solving it. Admittingly, I never would have solved it without the help of GPT, though even with that I had to figure a lot of this out for myself. I wish it could always spit out working code, but it only helps to get to the solution rather than giving the actual solution. Much of the issue was that the scaling values were just not being updated, and the code was overly complex for no good reason. It had 3 different variables referencing scale, and tried to scale everything inside the render target using a matrix. Now the RT is drawn at 1:1 scale and only scaled when drawn to the screen by referencing the user set variable "GameSettings.UiScale". At least this was part of the battle. The map code was also completely borked, and fixing the scale broke the map offsets when rescaling the UI (though to be fair, it was never correct to begin with). This was because the values that set everything up were never actually recalculated after a scaling event took place. It was clear the original dev(s) never intended the UI Scale to be available in the release version.