Move some of the settings around. Move screen shake to camera menu.
The "game" and "redux" settings are closely related but not identical, so move them closer together on the menus. To clarify the difference between the two, "Game" settings can slightly change some aspects of how the game is played but shouldn't change the core gameplay, while "Redux" settings can fully change aspects of the game and even the core gameplay. This also moves "Camera" further up the list under Redux, and moves the "Damage Screen Shake" option to the new camera menu. Language files were updated to reflect the new arrangement.