When viewing static images (other than photos) add 1 second delay before closing.
A short description can't really encapsulate a good summary so here's what it does. When viewing the Mr. Write "Christine" photo, the Lower Face Shrine mural, or the Alligator's painting in Animal Village, it is possible to close the dialog window immediately. This can create "accidental" closing of the image before getting a chance to view it. This commit adds a 1 second delay before the overlay can be closed.