Add a "FieldBarrier" around the current field in "Classic Camera" mode.
This massive change creates a new type of object called "ObjFIeldBarrier" and a new type of collision known as "Field" collision. When enabling the classic camera, the field barrier is established around the current field. This prevents objects and monsters outside of the field to interact with objects inside the field and vice versa. All monsters have been given the "Field" type collision, so they can never escape the current field. This simulates the single screen effect of the original games.

This change also fixes Link's jump/attack animations/interactions. It got wrapped up in this larger change, but at least it's fixed.