Project the sword beam into 2D space so it's behavior matches the original.
This makes it so as long as it looks like the sword shot hits an enemy, it will make a connection with it. So if you jump, shoot the sword, and the enemy is higher on the Y axis, the sword shot will make contact as the height from the Z value is added to the Y value. This doesn't make sense because physics, but it doesn't need to. It just needs to match the behavior of the original games.