Increase travel distance of sword beam and fix its position so it behaves "2D".
This adjusts the sword beam and makes it basically 100% "2D" when it comes to its collision. For example, if there is a wall in front of link with a small opening above it, he can jump, shoot the sword, and it will enter the opening instead of colliding with it. This also increases the distance of the sword shot so it at least travels a complete field.