Adjust the level 8 dungeon hacks to appear further from the front door.
There is two situations that manually set the map save position to the entrance to level 8: when entering one of the two back doors, or when using the ocarina to play Mambo's song. I'm not sure what happened, maybe current map changes, but teleporting to the location that was set forces the player to exit the dungeon when spawning there. So it's been adjusted slightly.