Classic Camera: Only update objects that are located within the current field.
This change makes it so that nothing is updated outside of the current field when classic camera is activated. This makes the game overall feel more like the original as in monsters will still remain in their spawn locations when first entering the screen. Of course it's not a perfect 1:1 recreation, but at least it's closer.