Add 2D "charge jumping" animations and fix attacking jumping.
This is the 2D version of ed7cbc4 which adds the new animations when charge jumping. Because of the sorry state of the 2D Link code, I had to fix other things too. Landing while charging set him back to idle state. This has been fixed. Attacking or charging then jumping reset the state so that's been fixed. Also landing when attacking or charging should keep the state so that's been fixed. Overall, I think 2D Link is finally feature complete.