Fix level 2 and level 6 dungeon map chest locations.
I made two stupid mistakes. The level 2 chest for the map tracks the variable "d2_dmap" and not "d2_map". The level 6 chest for the map was accidentally one tile lower than it should have been because the txt threw me off by having an extra "row" of numbers at the bottom that don't mean anything.