Classic Camera: Introduce camera snapping that actually works.
I attempted this before but the old system didn't work. When "smooth camera" is enabled during the classic camera, it is useful for screen transitions. It's not so useful when trying to change the scale or when transitioning between screens.

This system can be activated in two ways:
SnapCamera: This can be used when the start and end events are known like the map transition system. Set true to snap, set false to not.
SnapCameraTimer: This can be used when there is no clear start and end events like scaling. This disables smooth camera for a number of milliseconds.

The point of camera snapping is to get rid of the "jitter" when certain events take place. Like when changing the scale with the trigger buttons, or when changing maps during the transition system. This fixes both of these situations when "Smooth Camera" is enabled.