Fix two more animation states: jump > land while charging, jump > charge cancel.
Because there is never an end to the number of quirks with Link's animation states, we have this fix. This commits adds a hack to play the "stand" animation for 100ms shortly after landing while charging the sword so he doesn't appear to "slide" between jumps. The second fix resumes the jumping flip animation if cancelling a sword charge while in the air.