Classic Camera: Update followers when outside field + ObjLink refactoring.
When transitioning screens, followers may become temporarily outside the current field. This commit manually adds them to the corresponding update groups even when outside of the current field. This also updates ObjLink code so that similar elements are near each other. Most of the changes are in "Update3D" where parts of the code were branched into dedicated methods.