Classic Camera: Force constant calculated scale based on original dimensions.
The best way to go about a "Classic Camera" is to make it scale like one. With this commit, the scale is always automatically calculated using the original GameBoy dimensions (160x144) instead of the minimums used in this port (380x256). This allows for a much larger range of scaling values when resizing the window by almost double. For example, 3x scale before would be 1140x768 while after this change 3x is only 480x432. To match the old value with the new value, would require about a 6x scale (960x864). This change only affects the classic camera which no longer has the ability to zoom in and out, nor does it allow free camera.

This commit also removes the "Classic Blackout" option and just merges it with "Classic Border" so it's a single option. The 1 pixel border was never supposed to be an option to begin with, it was a testing option that I used to make sure the drawn border was always the correct size.