Classic Camera: Create a better "hide Link" transition hack.
When Link is transitioning between houses especially, he pops up at the top of the screen for a split second. This draws a rectangle that covers him up. This rectangle can't be seen because it is hidden by the circle shader that is drawn when transitioning. Why the circle shader does not cover this up by itself is a mystery. Because it is fully drawn over the screen during the entire transition. Well whatever, this fix works so meh.