More Controller/Direction shenanigans. Change "Classic Movement" behavior.
This simplifies the movement system by doing away with some of the "lahdmod" file options like "disable moonwalk" and "modern analog" when classic movement was enabled. The default behavior of Link when moving is now more akin to how "Classic Movement" used to work without the ability of walking backwards. Enabling it now just enables the ability to moonwalk.