Yet another animation hack. This time when jumping at end of an attack.
There is a window of a few frames, or maybe even frame perfect, where if attacking ends and a jump starts, neither the attacking state or the jumping state is set and the current state goes "Idle". This causes Link to stand in air when jumping. This hack attempts to catch the idle state at the beginning of the jump and instead transform it into the jumping state.