Classic Camera: Disable BowWow when in a dungeon.
This one kind of blew my mind... Apparently I still have no idea how this game fully works. A fix was implemented to keep BowWow active when he is a follower and the mode is classic camera. But I went inside the dungeon, and I could still him eating stuff... constantly and forever. But where is he? The overworld was not even loaded! None of the objects were being destroyed. BowWow was not in the map as far as I know! I have no idea. So on map init and in a dungeon, nullify the BowWow object.