Update several maps so that BowWow spawns in them.
I always thought it was weird that some maps had a "BowWow" face in them in the editor, and I never knew why. It probably should have been obvious as nothing was done that didn't have a purpose. Without the BowWow object he will not spawn on the map. So update several maps, especially caves, to have a BowWow spawner so he shows up on those maps.