Vacuum can't be killed with sword shot and fix direction getting stuck.
This was a weird one. The vacuum was killable with a sword shot, which then while it was doing its thing didn't actually stop applying a forced direction. When looking into it deeper, it seems that checking for it's current lives did not work for some reason. Maybe I am misunderstanding how "lives" works. Whatever the case, just end the rotation when it dies from any hit since it only has 1 HP anyway.