Classic Camera: Navigating the Egg.
This was really difficult to get right, and it's not perfect but close enough. I didn't want to have to resort to breaking the illusion that the rooms are "endless" until the correct path is taken. When classic camera is enabled, the logic is slightly different when going up and down, but the results are similar. I can't even begin to explain how the system works, the important thing is that it's working and the illusion is kept. It's possible to bug out the camera with just the right action for a split second, and there's not much I can do about that. When just playing normally, 99% of people will never trigger it.