Fix brief animation freeze when landing + performing certain actions.
There is a weird pause that can happen where Link's animation can freeze when landing if the following actions take place on the exact frame he hits the ground: sprinkling powder, throwing the boomerang, and poking with the sword. This commit detects those instances and prevents this from happening. This bug was the result of the "landing fix" hack forcing "stand" animation until it ends. I'm sure I'll find more actions that are bugged out in time...