When drowning move Link further into the water.
It looks like there was supposed to be code to do this but it never actually worked. It applied a velocity to the player which did nothing because it stopped them dead in their tracks. When walking into deep water without flipper or lava, move Link further out so it doesn't appear like he's sinking into land which looks weird.