Always animate Crow & Raven and despawn after some distance.
This commit adds Crows and Ravens to the always animate list, and also changes their behavior when they get a certain distance from the player. Before, they would fly forever until they eventually reached the end of the map. When moving around, you may have heard a random "flap" sound. That was the crow slightly entering the rendering zone and freezing up again.

When the distance is > 85 the flapping sound stops playing. When the distance is > 100 they fade out and despawn. Not sure why this wasn't handled as most enemies and objects have a despawn trigger after distance or time. Guess they were just overlooked.