Fix Spiny Beetle repelling while also vulnerable.
Not sure what is going on here. When lifting the object off of the beetle, it was acting as if the object was both there and wasn't. The beetle could be attacked but at the same time it would repel the sword attack and create a spark. From what I could tell the carriable component has a "IsPickedUp" field that for whatever reason is not going true. I followed the logic and it's working all the way up to the point the damage logic is checked. Well whatever, I implemented custom logic to track it and all is fine now.