Fix size of rock spawner field and prevent spawning when playing ocarina.
The size of the rock spawner field is determined by its object on the map editor. Turns out the original dev(s) were cheating and using a single spawner to cover two fields. This can't be done with classic camera, so two spawners was needed. But this meant it was extending into the field under the egg.