Revert Crow/Raven change (7f1d231) & update them to have starting delay.
The last "fix" didn't work like I thought it did, and the sound effect was still playing when the overworld loaded. So I dug into what was really happening. It seems for a brief period between map transitions, Link's position is set to 0,0 which triggers the Raven near the Dungeon 8 entrance. There was probably a few ways I could fix this but I just decided to give Crows and Ravens a 1 second idle period before they start searching for targets. This should be more than enough to prevent them from triggering until the transition finishes, and not enough to drastically change their behavior.