Revamp instant pickup objects (again) except make it work this time.
The last update to this "instant pickup" code worked decent enough, but the button still had to be held for at least 2 frames and didn't work so well with Spiny Beetles. These things have caused me so much grief. Just the slightest bit of interruption would cause the pickup to fail, the object to float in air, and the beetle to go invincible. So I reverted the last change, and decided to break the function apart, piece it back together, all while heavily documenting and modifying it.

The result is much better. A tap of the button now picks up all "instant pickup" objects including the Spiny Beetles. A pickup can not be interrupted once it starts, and the full pickup animation is played. Once an instant pickup is started, it stores the object and a bool which flags that the pickup is taking place. This is done to keep the "button held" until the object is lifted, and remembers the object for concurrent iterations until the object is lifted. This fixes all the issues with "instant pickup" objects and feels much better to use.