Disable Raven damage field component on map load and enable after delay.
The "always animate" stuff has exposed so many flaws with the core systems, and this is yet another one. When exiting a cave on Tal Tal Mountain Range, a Raven in the top left corner of the map was dealing damage immediately after exiting the cave. This is due to them being "always animated" and the systems temporarily moving Link somewhere near the top left of the map temporarily until it places him where he's supposed to be.

To be more far more specific to what is happening, when transitioning from a map to the overworld, Link's position on the Overworld temporarily uses the coordinates from the previous map. Just long enough to trigger the Ravens up there. Most maps are small, so most maps when exiting will put Link temporarily at the top left corner of the overworld. This glaring flaw should be addressed eventually, but I don't know if it can be. For now, the path of least resistance is to just neuter the Ravens for the split second Link appears up there.