Implement option menu tooltips.
Adds "PageTooltip.cs" to "Ingame\Pages". This is a feature I've been wanting to implement for months, it just took a long time to get here. There was a time when I would have had no idea how to even start, let alone make it as it is here. While I wrote all the English tooltips, the rest of the languages are machine translated. If someone native to those languages wants to make them better, than be my guest.