Fix some "always animate" objects hanging around after death (again).
This is a regression fix. Apparently where the temporary list is cleared is important. I moved the clear to just before the list is used, but the game didn't like that. Fireballs were once again crashing with "Map is null". Moving the clear call back to after the loop fixes it for some reason.