Classic Camera: Reset more objects when transitioning fields.
Resetting Beamos and their projectiles was a nightmare scenario. I had to link the beam to its host and vice versa and keep track of each "shot". A Beamos lazer can consist of up to 16 "balls" that make up the totality of the shot. So each one of these has to be deleted to fully remove the lazer. The card enemies were another treat to try to reset, along with a few others. I doubt this is the last of the "reset" stuff I'll be doing, but it knocks out a good chunk of it.