Classic Camera: Cover up transitional weirdness for "Bridge" map.
Classic Camera doesn't handle some transitions very well, as you can see parts of the transition in places you shouldn't. For example, on the bridge map, there is a very brief peek of Link entering the map on the left side of the screen. To cover that up, draw a rectangle over the transition. I hate these kinds of hacks, but there isn't much else I can do...