Fix occasional "hitch" when falling into 2D maps.
Sometimes when falling into a 2D map from a hole (like around Moldorm in Dungeon 1), Link will "hitch" in the air. This appears as something temporarily breaking his fall before hitting the ground. As if his fall speed suddenly stopped then picked up again. This was due to the "variable jump height" variables not properly getting reset when falling from above. So this commit fixes that.