InputHandler: Implement keyboard keys that make sense.
When using the editor, the keyboard doesn't just work, it depends on how keys are defined in InputHandler. For some reason, the keys were all over the place, and didn't match up with what would be expected. It's probably because the creators had an alternate keyboard layout or something. Whatever the case, I'm the one working on it now so I need the keyboard to behave as expected.