Classic Camera: Create a stone/pot respawner and respawn on field change.
In the original game, leaving the field and coming back respawns rocks and pots (as well as bushes and crystals but these are already handled). So this commit also implements respawning of pots. There was already a "ObjStoneSpawner" but this does not work in the same way and is disabled when Classic Camera is enabled.