Dungeon 6: Create another condition to open door above 3 Wizzrobes.
When entering the dungeon and taking a left, there is a room with 3 Wizzrobes. Defeating them opens the door to a room with a chest, and a door to the north. While you can not enter that north door, you may end up exiting through it later down the line. At that point the Wizzrobes will have respawned and the door will be closed. In the original game, this door would be open, so create a condition that opens the door when entering that room from above. When exiting the door to the Wizzrobe room, it will then close back up like the original game.