Adjust 2D jump height formulation for more granular control.
This commit gives a bit more control over 2D jump height. Before it was only possible to hold the button just long enough to get 2 maybe 3 different heights. With the new method there is a much greater possible range (closer to 5-7 heights). The difference is minor since jumps aren't all that fancy to begin with and it's hard to calculate a predetermined jump height based on "feel" with how quick the jumps are. At the very least the ability to have greater jump control is now there.