Differentiate between Gels spawned from Red Zol and ones already on map.
I was so happy to initially get this working I forgot to test if it affected Gels already on the map. And well, it did do something... it removed them completely upon a screen reset. The logic was not attempting to differentiate between a Gel already on the map and a Gel spawned from a Red Zol. So let's fix that.