Marin can swim in deep water when a follower.
This gives Marin sprites, animations, and the logic to swim in the water when Link has the flippers. This can not be achieved during normal gameplay as Marin is only with Link up until the point of getting the key to Level 4. This is only possible at the end of the game after finishing Level 8 and "Nothing is Missable" is enabled where Marin will ask to join Link. Because it is possible to take Marin into deep water at this point, I figured it would be better to just give her the ability to swim than have her walk on water.