Classic Camera: Implement Super Game Boy border "Classic Border" option.
When playing on some Game Boy Color emulators, it's possible they also support the Super Game Boy border despite this being a feature only available on the SNES and playing the game with only 4 colors. For example, loading Link's Awakening DX on a real SGB would give the cool border, but not the colors of the Game Boy color. Either way, I always thought that this was a cool feature that only an emulator can provide, and I wanted it here too.

This commit changes "Classic Camera" option "Classic Border" into a slider with three values: None, Black, and Super Game Boy. The border in this port differs slightly from the original in that Link's Awakening did not use the full 160x144 pixels, it only used 160x128 with 16 vertical pixels dedicated to the HUD. So I edited the border to add 16 additional rows of pixels to the screen area. The reason this is only available for Classic Camera is that it requires the black border and an integer scale. Normal camera can go outside of integer scales.

Last, this commit updates the migration tool and the patcher to support the border, which is created from other borders. This is borderline I know, but it's the best I can do to get it implemented discretely. This feature has been something I planned for a long time, but I've only really gotten to the point of being able to implement it recently. My knowledge of developing this has expanded greatly since I started, so doing it now was almost trivial. Anyway, hope you enjoy it, it's a fun way to experience the game.