Implement "Classic Music Cues" audio option.
This option slightly changes when overworld music plays based on the player's position in the overworld. When disabled, the environmental tilemap is more closely respected, and when enabled, the original game fields are more closely respected. For example, when disabled, far east Mabe Village continues to play the "town" music until lifting the rock and stepping into Ukuku Prairie. When enabled, the music change happens as soon as the field with the bushes and rocks is entered.