Rework Blade Traps to use a predictive movement system.
Rather than try to detect collision, the trap will instead predict how far it can move. If the distance detected is not enough to move, then movement is avoided completely. This prevents the trap from bashing against the block in places like Dungeon 1 after pushing a block in front of it.