Prevent movement during the alligator trade sequence.
These things are such a pain. Half of the logic is in script.zScript, the other half of the logic is on the object itself. So the solution always ends up being somewhere in between. In this case, when the trade is started, it sets a "alligator_freeze" value in "script.zScript" which the listener on "ObjAlligator" picks up and freezes Link during the trade.