Classic Camera: Don't respawn stones created from object respawners.
Some objects are created via "object respawners" that continue to respawn the object until it is no longer needed. An example is a pot inside the Level 3 entrance. Until the door is opened, the pot respawns. This commit just prevents a classic camera field transition from respawning objects created by object spawners which creates stacked duplicates of the object.