Centralize max game scale in "Game1" and sliders can jump by 5/10 with top buttons.
This commit adds a "max game scale" variable that controls the maximum value everywhere else in code. This makes it easier to change it to whatever whenever and not have to worry about updating it everywhere else. For example, if we want a maximum game scale of 50x, it's now as simple as "MaxGameScale = 50" and nothing else needs updated.

This also updates sliders so they can move in increments of 5x using LB/RB buttons or 10x using LT/RT buttons. For example, if step size is 1, then LB/RB would move the slider in steps of 5. If step size is 0.10, then it would move the slider in steps of 0.50. Eventually a very large maximum game scale is most likely going to be default so this will help navigate the sizes much easier.