Implement adjusting max game scale via lahdmodfile.
This commit makes it possible to adjust the maximum game scale via "Game1.lahdmod". It also adjusts holding the LT or RT buttons in-game to increase the "zoom" speed the longer the button is held. Finally, editor mode can also be enabled via Game1.lahdmod as opposed to using the command line.