Pegasus Boots dash can smash crystals with no sword equipped.
This commit makes it possible to run through large crystals with the Pegasus Boots when the sword is not equipped. The implementation was somewhat complex. Destroying the crystal requires detecting a "hit". The only time a hit registers when dashing is when the sword is out and makes contact, or when Link's body smashes into something and bounces back.

This implementation takes a different approach and creates a box in front of Link that appears when running, that is used to create a new hit type "CrystalSmash". The crystal object itself was made vulnerable to this new hit type. Because the box is slightly in front of Link, it smashes the crystal before Link's body makes contact with the crystal's collision box. This allows him to continue dashing through multiple crystals unimpeded.